UPDATED MainMenu.java: removed button data members, added functions instead. The compiler automatically creates and assigns listeners, all you need to do is set the buttons onClick attribute to the function name. I set all the buttons in the main menu already, to get your button working, just use the same code I did for my lights out button.
CREATED a lot of stuff, mostly finish my lights out view and controller. The model is incomplete as I'm still working on logic, I have a mostly operational model done at SU though.

FUTURE IDEAS: we should pass the view component that holds all the relevant info to our puzzle models instead of passing a string object that we then have to parse. I'm not sure if we can pass const View objects but that would be nice. We do still need to return a string though in our find solution function because the data model should not be updating the view